Moved content out of `examples`
