+++ contracts/dao/DAO.sol
@@ -4,7 +4,7 @@ import "@openzeppelin/contracts/math/SafeMath.sol";
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import "@openzeppelin/contracts/utils/EnumerableSet.sol";
 import "./VentureEth.sol";
-import "../../voting/Democratic.sol";
+import "../voting/Democratic.sol";
 
 
 /**
