+++ contracts/test/introspection/erc165/ERC20WhitelistedMock.sol
@@ -1,6 +1,6 @@
 pragma solidity ^0.6.0;
 import "@openzeppelin/contracts/access/Ownable.sol";
-import "./../../../../examples/introspection/erc165/ERC20Whitelisted.sol";
+import "./../../../introspection/erc165/ERC20Whitelisted.sol";
 
 
 contract ERC20WhitelistedMock is ERC20Whitelisted, Ownable {
